Adds signal handling and copy timeout adjustments
A copy timeout flag is added to set a custom value and support for
volumes <1G is added. Previously if you had a pvc that was, say, 512Mi,
the duration would be calculated as 0 * time.Second.

A parent context is created for the root command that will be done
on os.Kill and os.Interrupt to make the tool more polite and improve
cleanup. It is propagated into BaseStrategy to be used with client-go
calls.

A few minor improvements to quiet gopls analyses and warnings, moved
to an opts struct for strategies.NewBaseStrategy.